// contracts/generated/typescript/models/AuditEvent.ts
export interface AuditEvent {
    /**
     *
     * @type {string}
     * @memberof AuditEvent
     */
    id: string;
    /**
     *
     * @type {string}
     * @memberof AuditEvent
     */
    actorId: string;
    /**
     *
     * @type {string}
     * @memberof AuditEvent
     */
    actorDisplayName: string;
    /**
     *
     * @type {string}
     * @memberof AuditEvent
     */
    eventType: string;
    /**
     *
     * @type {string}
     * @memberof AuditEvent
     */
    aggregateType: string;
    /**
     *
     * @type {string}
     * @memberof AuditEvent
     */
    aggregateId: string;
    /**
     *
     * @type {string}
     * @memberof AuditEvent
     */
    changeBatchId?: string;
    /**
     *
     * @type {{ [key: string]: any; }}
     * @memberof AuditEvent
     */
    payload: { [key: string]: any; };
    /**
     *
     * @type {Array<ChangeTag>}
     * @memberof AuditEvent
     */
    tags: Array<ChangeTag>;
    /**
     *
     * @type {Date}
     * @memberof AuditEvent
     */
    createdAt: Date;
}